# docs/swagger.yaml
definitions:
  handler.CreateOpportunityRequest:
    properties:
      company:
        type: string
      link:
        type: string
      location:
        type: string
      remote:
        type: boolean
      role:
        type: string
      salary:
        type: integer
    type: object
  handler.CreateOpportunityResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.OpportunityResponse'
      message:
        type: string
    type: object
  handler.DeleteOpportunityResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.OpportunityResponse'
      message:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      errorCode:
        type: integer
      message:
        type: string
    type: object
  schemas.OpportunityResponse:
    properties:
      company:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      link:
        type: string
      location:
        type: string
      remote:
        type: boolean
      role:
        type: string
      salary:
        type: integer
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
  /opportunity:
    delete:
      consumes:
      - application/json
      description: Delete a job opportunity
      parameters:
      - description: Opportunity identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DeleteOpportunityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete opportunity
      tags:
      - Opportunities
    post:
      consumes:
      - application/json
      description: Create a new job opportunity
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOpportunityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CreateOpportunityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create opportunity
      tags:
      - Opportunities
swagger: "2.0"